Forecast made with XGBoost
In this third version of the notebook, the forecast is done on daily adjust closing stock price of FOE. The model is trained with daily data from 2016 to 2018 and is evaluated on the test data set which consists of the daily data from 2019 onwards. The forecast made with XGBoost has an MSE = 0.23 which is unusually accurate which leads me to doubt the credibility of the code structure but as far as I can tell, there doesn't seem to be anything wrong. Take this forecast with a grain of salt